Remove unused code and update configuration
