+++ src/ProjectOrigin.WalletSystem.Server/Options/ServiceOptions.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace ProjectOrigin.WalletSystem.Server.Options;
@@ -5,5 +6,5 @@ namespace ProjectOrigin.WalletSystem.Server.Options;
 public class ServiceOptions
 {
     [Required(AllowEmptyStrings = false)]
-    public string EndpointAddress { get; set; } = string.Empty;
+    public required Uri EndpointAddress { get; set; }
 }
