+++ src/ProjectOrigin.WalletSystem.Server/Services/GRPC/WalletService.cs
@@ -21,7 +21,7 @@ namespace ProjectOrigin.WalletSystem.Server.Services.GRPC;
 [Authorize]
 public class WalletService : V1.WalletService.WalletServiceBase
 {
-    private readonly string _walletSystemAddress;
+    private readonly Uri _walletSystemAddress;
     private readonly IUnitOfWork _unitOfWork;
     private readonly IHDAlgorithm _hdAlgorithm;
     private readonly IBus _bus;
@@ -61,7 +61,7 @@ public class WalletService : V1.WalletService.WalletServiceBase
             WalletDepositEndpoint = new V1.WalletDepositEndpoint()
             {
                 Version = 1,
-                Endpoint = _walletSystemAddress,
+                Endpoint = _walletSystemAddress.ToString(),
                 PublicKey = ByteString.CopyFrom(endpoint.PublicKey.Export())
             }
         };
