+++ src/ProjectOrigin.WalletSystem.Server/Services/REST/v1/SlicesController.cs
@@ -35,7 +35,7 @@ public class SlicesController : ControllerBase
     /// <response code="400">Public key could not be decoded.</response>
     /// <response code="404">Receiver endpoint not found.</response>
     [HttpPost]
-    [Route("v1/slices/")]
+    [Route("v1/slices")]
     [Produces("application/json")]
     [ProducesResponseType(StatusCodes.Status202Accepted)]
     [ProducesResponseType(StatusCodes.Status400BadRequest)]
@@ -82,24 +82,66 @@ public class SlicesController : ControllerBase
 
 #region Records
 
+/// <summary>
+/// Request to receive a certificate-slice from another wallet.
+/// </summary>
 public record ReceiveRequest
 {
-
+    /// <summary>
+    /// The public key of the receiving wallet.
+    /// </summary>
     public required byte[] PublicKey { get; init; }
+
+    /// <summary>
+    /// The sub-position of the publicKey used on the slice on the registry.
     public required uint Position { get; init; }
+
+    /// <summary>
+    /// The id of the certificate.
+    /// </summary>
     public required FederatedStreamId CertificateId { get; init; }
+
+    /// <summary>
+    /// The quantity of the slice.
+    /// </summary>
     public required uint Quantity { get; init; }
+
+    /// <summary>
+    /// The random R used to generate the pedersen commitment with the quantitiy.
+    /// </summary>
     public required byte[] RandomR { get; init; }
+
+    /// <summary>
+    /// List of hashed attributes, their values and salts so the receiver can access the data.
+    /// </summary>
     public required IEnumerable<HashedAttribute> HashedAttributes { get; init; }
 }
 
+/// <summary>
+/// Hashed attribute with salt.
+/// </summary>
 public record HashedAttribute
 {
+
+    /// <summary>
+    /// The key of the attribute.
+    /// </summary>
     public required string Key { get; init; }
+
+    /// <summary>
+    /// The value of the attribute.
+    /// </summary>
     public required string Value { get; init; }
+
+    /// <summary>
+    /// The salt used to hash the attribute.
+    /// </summary>
     public required byte[] Salt { get; init; }
 }
 
+/// <summary>
+/// Response to receive a certificate-slice from another wallet.
+/// </summary>
 public record ReceiveResponse { }
 
 #endregion
