+++ src/ProjectOrigin.WalletSystem.Server/Services/REST/v1/TransfersController.cs
@@ -141,3 +141,78 @@ public class TransfersController : ControllerBase
         });
     }
 }
+
+#region Records
+
+/// <summary>
+/// A transfer record of a transfer of a part of a certificate to another wallet.
+/// </summary>
+public record Transfer()
+{
+    public required FederatedStreamId FederatedStreamId { get; init; }
+    public required string ReceiverId { get; init; }
+    public required long Quantity { get; init; }
+    public required long Start { get; init; }
+    public required long End { get; init; }
+    public required string GridArea { get; init; }
+}
+
+/// <summary>
+/// A request to transfer a certificate to another wallet.
+/// </summary>
+public record TransferRequest()
+{
+    /// <summary>
+    /// The federated stream id of the certificate to transfer.
+    /// </summary>
+    public required FederatedStreamId CertificateId { get; init; }
+
+    /// <summary>
+    /// The id of the wallet to transfer the certificate to.
+    /// </summary>
+    public required Guid ReceiverId { get; init; }
+
+    /// <summary>
+    /// The quantity of the certificate to transfer.
+    /// </summary>
+    public required uint Quantity { get; init; }
+
+    /// <summary>
+    /// List of hashed attributes to transfer with the certificate.
+    /// </summary>
+    public required string[] HashedAttributes { get; init; }
+}
+
+/// <summary>
+/// A response to a transfer request.
+/// </summary>
+public record TransferResponse()
+{
+    /// <summary>
+    /// The id of the transfer request.
+    /// </summary>
+    public required Guid TransferRequestId { get; init; }
+}
+
+/// <summary>
+/// A result of aggregated transfers.
+/// </summary>
+public record AggregatedTransfers()
+{
+    /// <summary>
+    /// The start of the aggregated period.
+    /// </summary>
+    public required long Start { get; init; }
+
+    /// <summary>
+    /// The end of the aggregated period.
+    /// </summary>
+    public required long End { get; init; }
+
+    /// <summary>
+    /// The quantity of the aggregated transfers.
+    /// </summary>
+    public required long Quantity { get; init; }
+}
+
+#endregion
